Initial Load
Channel.cs - Keeps track of irc channel information and lists of objects
related to monitoring them
ExceptionHandling - Needs to be expanded upon but basically we need to
have an exception handling/logging class to catch bugs
IrcDotNet.xml/dll -- This is a custom compiled version of the ircdotnet
dll that fixed a circularbuffer issue that it had when in #speedrunslive

Newtonsoft.Json.dll -- parser for JSON responses.
Program.cs - Contains the basic IRC handling for the bot as well as
basic command parsing.
TwitchAPIInterface.cs - Contains the code to handle updating twitch
channels and parsing the data out into the TwitchStuff object.
Twitchstuff.cs - this object contains basic twitch response information
for us to use later on.